+++ stroom-dashboard-client/src/main/java/stroom/dashboard/client/text/TextPresenter.java
@@ -36,6 +36,7 @@ import stroom.dashboard.shared.ComponentConfig;
 import stroom.dashboard.shared.ComponentSettings;
 import stroom.dashboard.shared.Field;
 import stroom.dashboard.shared.IndexConstants;
+import stroom.dashboard.shared.TableComponentSettings;
 import stroom.dashboard.shared.TextComponentSettings;
 import stroom.dispatch.client.ClientDispatchAsync;
 import stroom.editor.client.presenter.EditorPresenter;
@@ -55,6 +56,7 @@ import stroom.util.shared.DefaultLocation;
 import stroom.util.shared.EqualsUtil;
 import stroom.util.shared.Highlight;
 import stroom.util.shared.OffsetRange;
+import stroom.util.shared.Version;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -63,6 +65,9 @@ import java.util.Set;
 
 public class TextPresenter extends AbstractComponentPresenter<TextPresenter.TextView> implements TextUiHandlers {
     public static final ComponentType TYPE = new ComponentType(2, "text", "Text");
+
+    private static final Version CURRENT_MODEL_VERSION = new Version(6, 1, 26);
+
     private final Provider<EditorPresenter> rawPresenterProvider;
     private final Provider<HtmlPresenter> htmlPresenterProvider;
     private final ClientDispatchAsync dispatcher;
@@ -262,15 +267,23 @@ public class TextPresenter extends AbstractComponentPresenter<TextPresenter.Text
             if (tablePresenter != null) {
                 final List<TableRow> selection = tablePresenter.getSelectedRows();
                 if (selection != null && selection.size() == 1) {
+                    final Field streamIdField = chooseBestField(tablePresenter, textSettings.getStreamIdField());
+                    final Field partNoField = chooseBestField(tablePresenter, textSettings.getPartNoField());
+                    final Field recordNoField = chooseBestField(tablePresenter, textSettings.getRecordNoField());
+                    final Field lineFromField = chooseBestField(tablePresenter, textSettings.getLineFromField());
+                    final Field colFromField = chooseBestField(tablePresenter, textSettings.getColFromField());
+                    final Field lineToField = chooseBestField(tablePresenter, textSettings.getLineToField());
+                    final Field colToField = chooseBestField(tablePresenter, textSettings.getColToField());
+
                     // Just use the first row.
                     final TableRow selected = selection.get(0);
-                    currentStreamId = getLong(textSettings.getStreamIdField(), selected);
-                    currentPartNo = getLong(textSettings.getPartNoField(), selected);
-                    currentRecordNo = getLong(textSettings.getRecordNoField(), selected);
-                    final Long currentLineFrom = getLong(textSettings.getLineFromField(), selected);
-                    final Long currentColFrom = getLong(textSettings.getColFromField(), selected);
-                    final Long currentLineTo = getLong(textSettings.getLineToField(), selected);
-                    final Long currentColTo = getLong(textSettings.getColToField(), selected);
+                    currentStreamId = getLong(streamIdField, selected);
+                    currentPartNo = getLong(partNoField, selected);
+                    currentRecordNo = getLong(recordNoField, selected);
+                    final Long currentLineFrom = getLong(lineFromField, selected);
+                    final Long currentColFrom = getLong(colFromField, selected);
+                    final Long currentLineTo = getLong(lineToField, selected);
+                    final Long currentColTo = getLong(colToField, selected);
 
                     // Validate settings.
                     if (textSettings.getStreamIdField() == null) {
@@ -353,6 +366,26 @@ public class TextPresenter extends AbstractComponentPresenter<TextPresenter.Text
         }
     }
 
+    private Field chooseBestField(final TablePresenter tablePresenter, final Field field) {
+        if (field != null) {
+            final TableComponentSettings tableComponentSettings = tablePresenter.getSettings();
+            final List<Field> fields = tableComponentSettings.getFields();
+            // Try and choose by id.
+            for (final Field f : fields) {
+                if (f.getId() != null && f.getId().equals(field.getId())) {
+                    return f;
+                }
+            }
+            // Try and choose by name.
+            for (final Field f : fields) {
+                if (f.getName() != null && f.getName().equals(field.getName())) {
+                    return f;
+                }
+            }
+        }
+        return null;
+    }
+
     private Long getLong(final Field field, final TableRow row) {
         if (field != null && row != null) {
             return getLong(row.getText(field.getId()));
@@ -427,18 +460,23 @@ public class TextPresenter extends AbstractComponentPresenter<TextPresenter.Text
         super.read(componentConfig);
         textSettings = getSettings();
 
+        final Version version = Version.parse(textSettings.getModelVersion());
+        final boolean old = version.lt(CURRENT_MODEL_VERSION);
+
         // special field names have changed from EventId to __event_id__ so we need to deal
         // with those and replace them, also rebuild existing special fields just in case
         if (textSettings.getStreamIdField() == null
-                || IndexConstants.STREAM_ID.equals(textSettings.getStreamIdField().getName())
-                || textSettings.getStreamIdField().isSpecial()) {
+                || (old && IndexConstants.STREAM_ID.equals(textSettings.getStreamIdField().getName()))
+                || (old && textSettings.getStreamIdField().isSpecial())) {
             textSettings.setStreamIdField(TablePresenter.buildSpecialField(IndexConstants.STREAM_ID));
         }
         if (textSettings.getRecordNoField() == null
-                || IndexConstants.EVENT_ID.equals(textSettings.getStreamIdField().getName())
-                || textSettings.getRecordNoField().isSpecial()) {
+                || (old && IndexConstants.EVENT_ID.equals(textSettings.getRecordNoField().getName()))
+                || (old && textSettings.getRecordNoField().isSpecial())) {
             textSettings.setRecordNoField(TablePresenter.buildSpecialField(IndexConstants.EVENT_ID));
         }
+
+        textSettings.setModelVersion(CURRENT_MODEL_VERSION.toString());
     }
 
     @Override