+++ stroom-dashboard-shared/src/main/java/stroom/dashboard/shared/TextComponentSettings.java
@@ -24,7 +24,20 @@ import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlType;
 
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "text", propOrder = {"tableId", "streamIdField", "partNoField", "recordNoField", "lineFromField", "colFromField", "lineToField", "colToField", "pipeline", "showAsHtml", "showStepping"})
+@XmlType(name = "text", propOrder = {
+        "tableId",
+        "streamIdField",
+        "partNoField",
+        "recordNoField",
+        "lineFromField",
+        "colFromField",
+        "lineToField",
+        "colToField",
+        "pipeline",
+        "showAsHtml",
+        "showStepping",
+        "modelVersion"
+})
 public class TextComponentSettings extends ComponentSettings {
     private static final long serialVersionUID = -2530827581046882396L;
 
@@ -50,6 +63,8 @@ public class TextComponentSettings extends ComponentSettings {
     private boolean showAsHtml;
     @XmlElement(name = "showStepping")
     private boolean showStepping;
+    @XmlElement(name = "modelVersion")
+    private String modelVersion;
 
     public TextComponentSettings() {
         // Default constructor necessary for GWT serialisation.
@@ -141,5 +156,13 @@ public class TextComponentSettings extends ComponentSettings {
 
     public void setShowStepping(final boolean showStepping) {
         this.showStepping = showStepping;
+    }
+
+    public String getModelVersion() {
+        return modelVersion;
+    }
+
+    public void setModelVersion(final String modelVersion) {
+        this.modelVersion = modelVersion;
     }
 }